+++ src/neptune/metadata_containers/project.py
@@ -49,7 +49,7 @@ from neptune.internal.utils import (
 )
 from neptune.metadata_containers import MetadataContainer
 from neptune.metadata_containers.abstract import NeptuneObjectCallback
-from neptune.metadata_containers.tables import Table
+from neptune.tables import Table
 from neptune.metadata_containers.utils import prepare_nql_query
 from neptune.types.mode import Mode
 from neptune.typing import (
