+++ tests/e2e/management/test_management.py
@@ -53,7 +53,7 @@ from neptune.management.exceptions import (
     WorkspaceOrUserNotFound,
 )
 from neptune.management.internal.utils import normalize_project_name
-from neptune.metadata_containers.tables import Table
+from neptune.tables import Table
 from tests.e2e.base import (
     BaseE2ETest,
     fake,
