+++ tests/unit/neptune/new/client/abstract_tables_test.py
@@ -35,7 +35,7 @@ from neptune.internal.backends.api_model import (
     LeaderboardEntry,
 )
 from neptune.internal.backends.neptune_backend_mock import NeptuneBackendMock
-from neptune.metadata_containers.tables import (
+from neptune.tables import (
     Table,
     TableEntry,
 )
